[reggen,topgen] Move top-level DV generation from reggen to topgen
After lots of refactoring in this code, we can now split this up
properly and reggen no longer needs the "Top" class (which definitely
felt like it was in the wrong place!)
