+++ util/topgen/top_uvm_reg.sv.tpl
@@ -4,7 +4,8 @@
 
 // UVM registers auto-generated by `reggen` containing UVM definitions for the entire top-level
 <%!
-  from reggen import gen_dv
+  from topgen.gen_dv import sv_base_addr
+  from reggen.gen_dv import bcname, mcname, miname
 %>
 ##
 ## This template is used for chip-wide tests. It expects to be run with the
@@ -78,14 +79,14 @@ ${make_ral_pkg_window_class(dv_base_prefix, 'chip', window)}
         esc_if_name = block.name.lower() + if_suffix
         if_inst = inst_name + if_suffix
 %>\
-    rand ${gen_dv.bcname(esc_if_name)} ${if_inst};
+    rand ${bcname(esc_if_name)} ${if_inst};
 %     endfor
 %   endfor
 % endfor
 % if windows:
     // memories
 %   for window in windows:
-    rand ${gen_dv.mcname('chip', window)} ${gen_dv.miname(window)};
+    rand ${mcname('chip', window)} ${miname(window)};
 %   endfor
 % endif
 
@@ -123,16 +124,16 @@ ${make_ral_pkg_window_class(dv_base_prefix, 'chip', window)}
           hdl_path = 'tb.dut.top_earlgrey.u_' + inst_name
         qual_if_name = (inst_name, if_name)
         base_addr = top.if_addrs[qual_if_name]
-        sv_base_addr = gen_dv.sv_base_addr(top, qual_if_name)
+        base_addr_txt = sv_base_addr(top, qual_if_name)
 %>\
-      ${if_inst} = ${gen_dv.bcname(esc_if_name)}::type_id::create("${if_inst}");
+      ${if_inst} = ${bcname(esc_if_name)}::type_id::create("${if_inst}");
       ${if_inst}.configure(.parent(this));
-      ${if_inst}.build(.base_addr(base_addr + ${sv_base_addr}), .csr_excl(csr_excl));
+      ${if_inst}.build(.base_addr(base_addr + ${base_addr_txt}), .csr_excl(csr_excl));
       ${if_inst}.set_hdl_path_root("${hdl_path}", "BkdrRegPathRtl");
       ${if_inst}.set_hdl_path_root("${hdl_path}", "BkdrRegPathRtlCommitted");
       ${if_inst}.set_hdl_path_root("${hdl_path}", "BkdrRegPathRtlShadow");
       default_map.add_submap(.child_map(${if_inst}.default_map),
-                             .offset(base_addr + ${sv_base_addr}));
+                             .offset(base_addr + ${base_addr_txt}));
 %     endfor
 %   endfor
 % endfor
